+++ R/prop_diff.R
@@ -51,7 +51,9 @@ NULL
 #'   non-responses.
 #'
 #' @return
-#' * `s_proportion_diff()` returns a named list of elements `diff` and `diff_ci`.
+#' * `s_proportion_diff()` returns a named list containing the elements `diff`,
+#'   `diff_ci`, and `diff_est_ci` (`diff_est_ci` combines the `diff` and `diff_ci`
+#'   values into a 3-element vector).
 #'   Depending on the method used, also the standard error of the difference `se_diff` is
 #'   returned.
 #'
@@ -112,7 +114,7 @@ s_proportion_diff <- function(df,
   method <- match.arg(method)
 
   if (is.null(.in_ref_col) || .in_ref_col) {
-    y <- list(diff = numeric(), diff_ci = numeric())
+    y <- list(diff = numeric(), diff_ci = numeric(), diff_est_ci = numeric())
   } else {
     checkmate::assert_false(is.null(.ref_group))
     assert_stratification_compatibility(
@@ -155,15 +157,17 @@ s_proportion_diff <- function(df,
 
     y$diff <- setNames(y$diff * 100, paste0("diff_", method))
     y$diff_ci <- setNames(y$diff_ci * 100, paste0("diff_ci_", method, c("_l", "_u")))
+    y$diff_est_ci <- c(y$diff, y$diff_ci)
     if (!is.null(y$se_diff)) {
       y$se_diff <- setNames(y$se_diff * 100, paste0("se_diff_", method))
     }
   }
 
   attr(y$diff, "label") <- "Difference in Response rate (%)"
-  attr(y$diff_ci, "label") <- d_proportion_diff(conf_level, method, long = FALSE)
+  attr(y$diff_ci, "label") <- tern::d_proportion_diff(conf_level, method, long = FALSE)
+  attr(y$diff_est_ci, "label") <- paste(attr(y$diff, "label"), "and", attr(y$diff_ci, "label"))
   if (!is.null(y$se_diff)) {
-    attr(y$se_diff, "label") <- paste0("Standard Error of Difference in Response rate (%)")
+    attr(y$se_diff, "label") <- paste("Standard Error of", attr(y$diff, "label"))
   }
 
   y
@@ -308,7 +312,7 @@ estimate_proportion_diff <- function(lyt,
                                      .stat_names = NULL,
                                      .formats = c(diff = "xx.x", diff_ci = "(xx.x, xx.x)", se_diff = "xx.x"),
                                      .labels = NULL,
-                                     .indent_mods = c(diff = 0L, diff_ci = 1L, se_diff = 1L)) {
+                                     .indent_mods = c(diff = 0L, diff_est_ci = 0L, diff_ci = 1L, se_diff = 1L)) {
   # Depending on main functions
   extra_args <- list(
     "na_rm" = na_rm,